ev-view: Lazily create annotation windows
`EvView` creates a window for each annotation when the page loads, which
isn't really necessary as the window will be created when the annotation
gets clicked. As such, we can remove this creation on page load so
evince only creates annotation windows when necessary.

This (parially) fixes large memory usage of Evince when loading a
document with a lot of annotations. The problem seemed to be that
`EvAnnotationWindow` loads a `GSpellTextView` with its own
`HunspellChecker` instance, creating a lot of unnecessary overhead.

Partially fixes https://gitlab.gnome.org/GNOME/evince/issues/1010